requirements.txt
torch==2.0.1
transformers==4.30.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
beautifulsoup4==4.12.2
scrapy==2.9.0
requests==2.31.0
rdkit==2023.3.2
gradio==3.35.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
jupyter==1.0.0
pytest==7.4.0
pyyaml==6.0
tqdm==4.65.0
wandb==0.15.4
tensorboard==2.13.0
